+++ core/primitives-core/src/lib.rs
@@ -8,8 +8,10 @@ pub mod apply;
 pub mod chains;
 pub mod code;
 pub mod config;
+pub mod deterministic_account_id;
 pub mod errors;
 pub mod gas;
+pub mod global_contract;
 pub mod hash;
 pub mod serialize;
 pub mod types;
